[FLINK-5734] remove cache mechanism from TemplateManager
